done so far with auth

+++ part4/controllers/blogs.js
@@ -1,18 +1,16 @@
 const blogRouter = require("express").Router();
 const Blog = require("../models/blogModel");
 const User = require("../models/userModel");
-const jwt = require('jsonwebtoken');
+const jwt = require("jsonwebtoken");
 //testing
 const blogList = require("../utils/list_helper");
 //app.use("/api/blogs", blogRoutes); esta linea ki que hace es poner /api/blogs como prefijo de cualquiera
 // de nuestras rutas
 const tokenExtractorMiddleware = require("../utils/middleware");
 
-
-
 blogRouter.get("/", async (req, res, next) => {
   try {
-    const allBlogs = await Blog.find({}).populate("user", {username: 1});
+    const allBlogs = await Blog.find({}).populate("user", { username: 1 });
     if (allBlogs) {
       res.json(allBlogs);
     } else {
@@ -31,8 +29,7 @@ blogRouter.post("/", tokenExtractorMiddleware, async (req, res, next) => {
   const decodeToken = jwt.verify(req.token, process.env.SECRET);
 
   if (!decodeToken.id) {
-    return res.status(401).send({error: "token invalid"});
-     
+    return res.status(401).send({ error: "token invalid" });
   }
 
   const user = await User.findById(decodeToken.id);
@@ -40,7 +37,6 @@ blogRouter.post("/", tokenExtractorMiddleware, async (req, res, next) => {
   try {
     if (!req.body.title || !req.body.url) {
       return res.status(400).end();
-      
     }
     const blog = new Blog({
       title,
@@ -56,21 +52,28 @@ blogRouter.post("/", tokenExtractorMiddleware, async (req, res, next) => {
     user.blogs = user.blogs?.concat(savedBlog._id);
     await user.save();
 
-  return   res.status(201).json(savedBlog);
- // res.json(savedBlog)
+    return res.status(201).json(savedBlog);
   } catch (error) {
     next(error);
   }
 });
 
-
-blogRouter.put("/:id",tokenExtractorMiddleware, async (req, res, next) => {
+blogRouter.put("/:id", tokenExtractorMiddleware, async (req, res, next) => {
   //update just the likes of the app
+  //finding the user 
+  const decodeToken = jwt.verify(req.token, process.env.SECRET);
+  if (!decodeToken.id) {
+    return res.status(401).send({ error: "token invalid" });
+  }
+  const user = await User.findById(decodeToken.id);
   try {
+
     const { likes } = req.body;
     const updatedBlogToSend = {
+      //añadir ref al user once front-end is done
       ...req.body,
       likes: likes,
+      user: user._id,
     };
     //update
     const updatedBlog = await Blog.findByIdAndUpdate(
@@ -88,7 +91,6 @@ blogRouter.put("/:id",tokenExtractorMiddleware, async (req, res, next) => {
   }
 });
 
-
 blogRouter.get("/:id", async (req, res, next) => {
   try {
     const idToSearch = req.params.id;
@@ -102,22 +104,24 @@ blogRouter.get("/:id", async (req, res, next) => {
   }
 });
 
-
 blogRouter.delete("/:id", tokenExtractorMiddleware, async (req, res) => {
+  
+  const decodeToken = jwt.verify(req.token, process.env.SECRET);
+  if (!decodeToken) {
+    return res.status(401).send({ error: "Unauthorized to delete this" });
+  }
   try {
-
     const singleblog = await Blog.findById(req.params.id);
-
-    const decodeToken = jwt.verify(req.token, process.env.SECRET);
 
     if (singleblog.user.toString() === decodeToken.id) {
       const author = await Blog.findByIdAndDelete(req.params.id);
       res.status(200).json(author);
-    } else {
-      res.status(401).send({
-        error: "Unauthorized to delete this",
-      });
     }
+    //else {
+    //   res.status(401).send({
+    //     error: "Unauthorized to delete this",
+    //   });
+    // }
 
     //const author = await Blog.findByIdAndDelete(req.params.id);
   } catch (err) {
